feat: Move code analysis to server-side function
Refactor the CodeVisualizer to use a Supabase Edge Function for all heavy processing, making the client-side application more responsive and scalable.

This commit introduces a new Supabase Edge Function `analyze-codebase` that is responsible for:
1.  Fetching and analyzing code from a GitHub repository URL.
2.  Processing and analyzing code from user-uploaded files.

The `CodeVisualizer.tsx` component has been refactored to remove all client-side analysis logic (including the web worker implementation) and now calls this new edge function.

This change moves the performance-intensive work from the user's browser to the backend, resulting in a significantly better user experience and a more robust architecture.
